Improve Privacy of `tetris_game.h`
Move information only used in `tetris_game.c` into a file:
`tetris_game_priv.h`
